requirements.txt
https://github.com/wroberts/pytimeparse.git
https://github.com/Rapptz/discord-ext-menus.git
discord.py
